Revert "Revert "Trying to figure out z-buffering issue""
This reverts commit 5b14c3b4ee41fd6c40cb310bf83d3ba8ce6caf03.

+++ GaleEngine/Rendering/Texture.cpp
@@ -19,7 +19,7 @@ void Texture::Destroy() {
 
 Texture::Texture(const string &filename) {
 	name = filename;
-	unsigned int dotPos = filename.find('.');
+	unsigned int dotPos = (unsigned int)filename.find('.');
 	string extension = filename.substr(dotPos);
 
 	if (extension == ".bmp" || extension == ".BMP") {
